+++ internal/core/user.go
@@ -1,10 +1,10 @@
 package core
 
-func (s *Service) CreateUserFromTG(login, password string, TGID int) (id string, err error) {
+func (s *Service) CreateUserFromTG(login, password string, TGID int64) (id string, err error) {
 	return s.dbManager.CreateNewUserTG(login, password, TGID)
 }
 
-func (s *Service) ConnectUserToTG(userID string, telegramID int) (err error) {
+func (s *Service) ConnectUserToTG(userID string, telegramID int64) (err error) {
 	err = s.dbManager.UpdateUserTGID(userID, telegramID)
 	if err != nil {
 		return err
@@ -12,7 +12,7 @@ func (s *Service) ConnectUserToTG(userID string, telegramID int) (err error) {
 	return nil
 }
 
-func (s *Service) ValidateUserTG(telegramID int) (id string, err error) {
+func (s *Service) ValidateUserTG(telegramID int64) (id string, err error) {
 	//find user by telegram ID
 	userID, err := s.dbManager.GetUserByTGID(telegramID)
 	if err != nil {
